+++ app/Http/Controllers/ExamQuestionController.php
@@ -9,12 +9,18 @@ use App\Http\Requests\PostEditExamQuestionFormRequest;
 use App\Http\Controllers\Controller;
 use App\ExamQuestion;
 use App\ExamQuestionAnswer;
+use Gate;
 
 class ExamQuestionController extends Controller
 {
     public function getApi(Request $request)
     {
-        $exam_question = new ExamQuestion;
+        if (Gate::denies('read-exam-question'))
+        {
+            return abort(401);
+        }
+
+        $exam_question = ExamQuestion::with('exam.author');
 
         if ($request->has('exam_id'))
         {
@@ -26,6 +32,11 @@ class ExamQuestionController extends Controller
 
     public function getView($id)
     {
+        if (Gate::denies('read-exam-question'))
+        {
+            return abort(401);
+        }
+
         $exam_question = ExamQuestion::findOrFail($id);
 
         return view('exam.question.view', compact('exam_question'));
@@ -33,6 +44,11 @@ class ExamQuestionController extends Controller
 
     public function getEdit($id)
     {
+        if (Gate::denies('update-exam-question'))
+        {
+            return abort(401);
+        }
+
         $exam_question = ExamQuestion::findOrFail($id);
 
         return view('exam.question.edit', compact('exam_question'));
@@ -99,6 +115,11 @@ class ExamQuestionController extends Controller
 
         try
         {
+            if (Gate::denies('delete-exam-question'))
+            {
+                throw new \Exception(trans('error.unauthorized.action'));
+            }
+
             ExamQuestion::findOrFail($id)->delete();
             ExamQuestionAnswer::where('exam_question_id', $id)->delete();
